+++ packages/cockpit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @creezio/cockpit
 
+## 0.20.0
+
+### Patch Changes
+
+- @creezio/shell-ui@0.20.0
+
 ## 0.19.0
 
 ### Patch Changes